+++ amethyst_input/src/util.rs
@@ -1,25 +1,29 @@
 use winit::{ElementState, Event, KeyboardInput, VirtualKeyCode, WindowEvent};
 
-pub fn get_key(event: &Event) -> Option<VirtualKeyCode> {
+pub fn get_key(event: &Event) -> Option<(VirtualKeyCode,ElementState)> {
     match *event {
         Event::WindowEvent { ref event, .. } => match *event {
             WindowEvent::KeyboardInput {
                 input:
                     KeyboardInput {
                         virtual_keycode: Some(ref virtual_keycode),
-                        state: ElementState::Released,
+                        state,
                         ..
                     },
                 ..
-            } => Some(*virtual_keycode),
+            } => Some((*virtual_keycode,state)),
             _ => None,
         },
         _ => None,
     }
 }
 
-pub fn is_key(event: &Event, key_code: VirtualKeyCode) -> bool {
-    get_key(event) == Some(key_code)
+pub fn is_key_down(event: &Event, key_code: VirtualKeyCode) -> bool {
+    let op = get_key(event);
+    if let Some((key,state)) = op {
+        return key == key_code && state == ElementState::Pressed;
+    }
+    return false;
 }
 
 pub fn is_close_requested(event: &Event) -> bool {