+++ docs/CHANGELOG.md
@@ -47,6 +47,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 * Point lights use `GlobalTransform` for positioning rather than a separate `center` ([#794])
 * Point lights now require a `GlobalTransform` component to be included in rendering ([#794])
 * `amethyst_input::input_handler::{keys_that_are_down, mouse_buttons_that_are_down, scan_codes_that_are_down, buttons_that_are_down}` now all return `impl Iterator` instead of concrete wrapper types ([#816])
+* Renamed is_key to is_key_down and fixed example to react when the key is pressed instead of released. ([#822])
 
 ### Removed
 * Remove `amethyst_input::{KeyCodes, ScanCodes, MouseButtons, Buttons}` in favor of `impl trait` ([#816])
@@ -91,6 +92,7 @@ The format is based on [Keep a Changelog][kc], and this project adheres to
 [#815]: https://github.com/amethyst/amethyst/pull/815
 [#817]: https://github.com/amethyst/amethyst/pull/817
 [#818]: https://github.com/amethyst/amethyst/pull/818
+[#822]: https://github.com/amethyst/amethyst/pull/822
 
 ## [0.7.0] - 2018-05
 ### Added